backend/requirements.txt
fastapi==0.104.0
uvicorn==0.22.0
sqlalchemy==2.0.29
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
numpy
opencv-python